+++ testcontainers/src/core/ports.rs
@@ -2,7 +2,7 @@ use bollard_stubs::models::{PortBinding, PortMap};
 use std::{collections::HashMap, net::IpAddr};
 
 /// The exposed ports of a running container.
-#[derive(Clone, Debug, Default, PartialEq)]
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
 pub struct Ports {
     ipv4_mapping: HashMap<u16, u16>,
     ipv6_mapping: HashMap<u16, u16>,
@@ -96,7 +96,7 @@ impl From<PortMap> for Ports {
 
 fn parse_port(port: &str) -> u16 {
     port.parse()
-        .unwrap_or_else(|e| panic!("Failed to parse {} as u16 because {}", port, e))
+        .unwrap_or_else(|e| panic!("Failed to parse {port} as u16 because {e}"))
 }
 
 #[cfg(test)]